Rename Callbacks project to Telephony (1 of 2)
This project is no longer just about receiving callbacks
from the telephony provider - it now encapsulates all interaction
with the telephony provider
