--- apps/callbacks/test/lib/callbacks/twiml_test.exs
@@ -1,4 +1,0 @@
-defmodule Callbacks.TwimlTest do
-  use ExUnit.Case, async: true
-  doctest Callbacks.Twiml
-end
