+++ apps/telephony/priv/gettext/en/LC_MESSAGES/errors.po
