+++ apps/telephony/priv/gettext/errors.pot
