+++ apps/telephony/priv/static/favicon.ico
